chore revamped the backend

+++ app/Http/Controllers/UserController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\CreateUserRequest;
 use App\Http\Requests\UpdateProfileRequest;
 use App\Http\Requests\UserRequest;
 use App\Models\User;
@@ -14,8 +15,12 @@ class UserController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function index()
+    public function index(Request $request)
     {
+        if ($request->user()->type != 'admin') {
+            return response()->json(["message" => "You are not authorized"], 401);
+        }
+
         $User = User::all();
         return response()->json(["data" => $User], 200);
     }
@@ -26,13 +31,33 @@ class UserController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function show($id)
+    public function show(Request $request, $id)
     {
+        if ($request->user()->type != 'admin') {
+            return response()->json(["message" => "You are not authorized"], 401);
+        }
+
         $User = User::findOrFail($id);
         return response()->json(["data" => $User], 200);
     }
 
-    // update user info
+    public function store(CreateUserRequest $request)
+    {
+        if ($request->user()->type != 'admin') {
+            return response()->json(["message" => "You are not authorized"], 401);
+        }
+        
+        $User = new User;
+        $User->full_name = $request->full_name;
+        $User->email = $request->email;
+        $User->phone_number = $request->phone_number;
+        $User->password = bcrypt($request->password);
+        $User->role_id = $request->role_id;
+        $User->type = 'cooperativemanager';
+        $User->save();
+
+        return response()->json(["data" => $User], 200);
+    }
 
     /**
      * Update the specified resource in storage.
@@ -43,12 +68,19 @@ class UserController extends Controller
      */
     public function update(UpdateProfileRequest $request, $id)
     {
+        if ($request->user()->id != $id && $request->user()->type != 'admin') {
+            return response()->json(["message" => "You are not authorized"], 401);
+        }
+
         $User = User::findOrFail($id);
-        if ($request->name) {
-            $User->name = $request->name;
+        if ($request->full_name) {
+            $User->full_name = $request->full_name;
         }
         if ($request->email) {
             $User->email = $request->email;
+        }
+        if ($request->phone_number) {
+            $User->phone_number = $request->phone_number;
         }
         if ($request->password) {
             $User->password = bcrypt($request->password);
@@ -64,8 +96,12 @@ class UserController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+    public function destroy(Request $request, $id)
     {
+        if ($request->user()->type != 'admin') {
+            return response()->json(["message" => "You are not authorized"], 401);
+        }
+
         $User = User::findOrFail($id);
         $User->delete();
